Delete item -> delete image file if exists

+++ routes/index.js
@@ -3469,11 +3469,14 @@ router.post('/item/delete', isAuthenticated, function (req, res ) {
 
         // borra la imagen anterior
         var img_path = path.join(__dirname, '..', 'uploads/', data[0].nombre_imagen);
-        fs.unlinkSync( img_path );
-        console.log('successfully deleted '+ img_path);
+
+        if ( fs.existSync( img_path )) {
+            fs.unlinkSync(img_path);
+            console.log('successfully deleted ' + img_path);
+        }
 
 
-        if ( data[1] != null ){
+        if ( data[1] !== null ){
             console.log('El proveedor saldo a cuenta del proveedor '+ data[1].nombre +' ha sido actualizado' );
         }
